[SPARK-31127][ML] Add abstract Selector
